Using fetch API to call service
- Unsing Fetch API
