+++ server/templates/index.html
@@ -39,6 +39,6 @@
             <span id='score'>0</span>
         </div>
     </body>
-    <script src="scripts/main.js"></script>
+    <script src="http://192.168.25.5:5000/static/scripts/main.js"></script>
     
 </html>
